history_skin

+++ utils/utils.py
@@ -3,7 +3,7 @@ import json
 import subprocess
 import sys
 import urllib.parse
-
+import os
 
 def generate_market_url(skin_name):
     """Генерирует URL для скина на маркете Steam."""
@@ -17,28 +17,43 @@ def check_skin_in_database(skin, skin_data):
         return True
     return False
 
+
+
 def save_data(new_data, filename):
-    """Сохраняет или обновляет данные в JSON файле."""
+    """Сохраняет или обновляет данные в JSON файле более эффективно."""
     try:
-        with open(filename, 'r', encoding='utf-8') as f:
-           existing_data = json.load(f)
-        if not existing_data:
-            existing_data = {}
-        for key, value in new_data.items():
-            if key in existing_data:
-                if isinstance(existing_data[key], list):
-                    existing_data[key].append(value)
-                elif isinstance(existing_data[key], dict):
-                    existing_data[key].update(value)
-                else:
-                    existing_data[key] = [existing_data[key], value]
-            else:
-                existing_data[key] = value
-        with open(filename, 'w', encoding='utf-8') as f:
-            json.dump(existing_data, f, ensure_ascii=False, indent=4)
+        if os.path.exists(filename) and os.path.getsize(filename) > 0:
+            with open(filename, 'r+', encoding='utf-8') as f:
+                try:
+                    existing_data = json.load(f)
+                except json.JSONDecodeError:
+                    existing_data = {}  # Если файл пустой или поврежден
+                if not existing_data:
+                    existing_data = {}
+
+                for key, value in new_data.items():
+                    if key in existing_data:
+                        if isinstance(existing_data[key], list):
+                            existing_data[key].append(value)
+                        elif isinstance(existing_data[key], dict):
+                            existing_data[key].update(value)
+                        else:
+                            existing_data[key] = [existing_data[key], value]
+                    else:
+                        existing_data[key] = value
+
+                # Перезаписываем файл без повторного открытия
+                f.seek(0)
+                json.dump(existing_data, f, ensure_ascii=False, indent=4)
+                f.truncate()
+        else:
+            with open(filename, 'w', encoding='utf-8') as f:
+                json.dump(new_data, f, ensure_ascii=False, indent=4)
+
         print(f"Данные успешно сохранены в {filename}")
     except Exception as e:
         print(f"Ошибка при сохранении данных: {e}")
+
         
     
 def signal_handler(signum, frame):
